give up on custom ticker, reset ticker in execute to back off
