+++ pkg/backoff/counter_test.go
@@ -59,49 +59,16 @@ func TestMultiplicativeCounter(t *testing.T) {
 		tt := tt
 		t.Run(tt.name, func(t *testing.T) {
 			t.Parallel()
-			ec := newMultiplicativeCounter(tt.baseInterval, tt.maxInterval)
+			ec := NewMultiplicativeDurationCounter(tt.baseInterval, tt.maxInterval)
 			for _, expected := range tt.expected {
-				require.Equal(t, expected, ec.next())
+				require.Equal(t, expected, ec.Next())
+			}
+
+			ec.Reset()
+
+			for _, expected := range tt.expected {
+				require.Equal(t, expected, ec.Next())
 			}
 		})
 	}
-}
-
-// TestMultiplicativeTicker tests the NewMultiplicativeTicker and its behavior.
-func TestMultiplicativeTicker(t *testing.T) {
-	t.Parallel()
-
-	baseTime := 100 * time.Millisecond
-	maxTime := 500 * time.Millisecond
-
-	tk := NewMultiplicativeTicker(baseTime, maxTime)
-	defer tk.Stop()
-
-	expectedDurations := []time.Duration{
-		100 * time.Millisecond,
-		200 * time.Millisecond,
-		300 * time.Millisecond,
-		400 * time.Millisecond,
-		500 * time.Millisecond, // maxTime limit
-		500 * time.Millisecond, // maxTime limit
-	}
-
-	buffer := 50 * time.Millisecond
-
-	for _, expected := range expectedDurations {
-		start := time.Now()
-
-		select {
-		case <-tk.C:
-			require.WithinDuration(t, start, time.Now(), expected+buffer)
-		case <-time.After(maxTime + buffer):
-			t.Fatalf("ticker did not send event in expected time: %v", expected)
-		}
-	}
-
-	// stop the ticker
-	tk.Stop()
-
-	// call stop again to make sure no panic (same as stdlib ticker)
-	tk.Stop()
 }
